button.js renamed
